# The-DoorGym-App/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym Workout App API - proxies Adalo database
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Workout {
  id: string;
  name: string;
  description?: string;
  image?: string | null;
  /** Duration in minutes */
  duration: number;
  difficulty?: string | null;
  category?: string | null;
  exerciseCount: number;
  calories?: number | null;
}

export interface Exercise {
  id: string;
  name: string;
  description: string;
  image?: string | null;
  muscleGroup?: string | null;
  sets?: number | null;
  reps?: string | null;
  duration?: number | null;
  equipment?: string | null;
  videoUrl?: string | null;
}

export interface Category {
  id: string;
  name: string;
  image?: string | null;
  workoutCount?: number | null;
}

export interface WorkoutListResponse {
  records: Workout[];
  offset: number;
  nextOffset?: number | null;
}

export interface ExerciseListResponse {
  records: Exercise[];
  offset: number;
  nextOffset?: number | null;
}

export interface CategoryListResponse {
  records: Category[];
  offset: number;
  nextOffset?: number | null;
}

export type GetWorkoutsParams = {
categoryId?: string;
limit?: number;
offset?: number;
};

export type GetExercisesParams = {
categoryId?: string;
muscleGroup?: string;
limit?: number;
offset?: number;
};