# The-DoorGym-App/lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym Workout App API - proxies Adalo database
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: string;
  name: string;
  image?: string | null;
  workoutCount?: number | null;
}