# The-DoorGym-App/lib/api-zod/src/generated/types/exercise.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym Workout App API - proxies Adalo database
 * OpenAPI spec version: 0.1.0
 */

export interface Exercise {
  id: string;
  name: string;
  description: string;
  image?: string | null;
  muscleGroup?: string | null;
  sets?: number | null;
  reps?: string | null;
  duration?: number | null;
  equipment?: string | null;
  videoUrl?: string | null;
}