# The-DoorGym-App/lib/api-zod/src/generated/types/exerciseListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym Workout App API - proxies Adalo database
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from './exercise';

export interface ExerciseListResponse {
  records: Exercise[];
  offset: number;
  nextOffset?: number | null;
}