# The-DoorGym-App/lib/api-zod/src/generated/types/workout.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym Workout App API - proxies Adalo database
 * OpenAPI spec version: 0.1.0
 */

export interface Workout {
  id: string;
  name: string;
  description?: string;
  image?: string | null;
  /** Duration in minutes */
  duration: number;
  difficulty?: string | null;
  category?: string | null;
  exerciseCount: number;
  calories?: number | null;
}