# The-DoorGym-App/lib/api-zod/src/generated/types/workoutListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gym Workout App API - proxies Adalo database
 * OpenAPI spec version: 0.1.0
 */
import type { Workout } from './workout';

export interface WorkoutListResponse {
  records: Workout[];
  offset: number;
  nextOffset?: number | null;
}